# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Examination System API",
    "description": "API documentation for the Examination System",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/api/courses": {
      "post": {
        "summary": "Create a new course (Admin only)",
        "tags": ["Courses"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "description"],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Web Development"
                  },
                  "description": {
                    "type": "string",
                    "example": "Learn modern web development techniques with HTML, CSS, and JavaScript"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Course created successfully"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Not authorized"
          },
          "403": {
            "description": "Not allowed to access this route"
          }
        }
      },
      "get": {
        "summary": "Get all courses (Admin sees all, Member sees assigned)",
        "tags": ["Courses"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of courses"
          },
          "401": {
            "description": "Not authorized"
          }
        }
      }
    },
    "/api/courses/{id}": {
      "get": {
        "summary": "Get single course",
        "tags": ["Courses"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string",
              "example": "60d0fe4f5311236168a109cb"
            },
            "required": true,
            "description": "Course ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Course data"
          },
          "401": {
            "description": "Not authorized"
          },
          "403": {
            "description": "Not allowed to access this course"
          },
          "404": {
            "description": "Course not found"
          }
        }
      }
    },
    "/api/courses/{id}/assign": {
      "post": {
        "summary": "Assign course to member (Admin only)",
        "tags": ["Courses"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string",
              "example": "60d0fe4f5311236168a109cb"
            },
            "required": true,
            "description": "Course ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["memberId"],
                "properties": {
                  "memberId": {
                    "type": "string",
                    "example": "60d0fe4f5311236168a109ca"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Course assigned successfully"
          },
          "400": {
            "description": "Invalid input or course already assigned"
          },
          "401": {
            "description": "Not authorized"
          },
          "403": {
            "description": "Not allowed to access this route"
          },
          "404": {
            "description": "Course or member not found"
          }
        }
      }
    },
...
    "/api/results": {
      "get": {
        "summary": "Get all exam results (Admin sees all, Member sees own)",
        "tags": ["Results"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of exam results"
          },
          "401": {
            "description": "Not authorized"
          }
        }
      }
    },
    "/api/results/{id}": {
      "get": {
        "summary": "Get single exam result",
        "tags": ["Results"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string",
              "example": "60d0fe4f5311236168a109cd"
            },
            "required": true,
            "description": "Result ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Exam result data"
          },
          "401": {
            "description": "Not authorized"
          },
          "403": {
            "description": "Not allowed to access this result"
          },
          "404": {
            "description": "Result not found"
          }
        }
      }
    },
    "/api/results/member/{memberId}": {
      "get": {
        "summary": "Get results for a specific member (Admin only)",
        "tags": ["Results"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "memberId",
            "schema": {
              "type": "string",
              "example": "60d0fe4f5311236168a109ca"
            },
            "required": true,
            "description": "Member ID"
          }
        ],
        "responses": {
          "200": {
            "description": "List of member's exam results"
          },
          "401": {
            "description": "Not authorized"
          },
          "403": {
            "description": "Not allowed to access this route"
          }
        }
      }
    },
    "/api/results/course/{courseId}": {
      "get": {
        "summary": "Get results for a specific course (Admin only)",
        "tags": ["Results"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "courseId",
            "schema": {
              "type": "string",
              "example": "60d0fe4f5311236168a109cb"
            },
            "required": true,
            "description": "Course ID"
          }
        ],
        "responses": {
          "200": {
            "description": "List of course exam results"
          },
          "401": {
            "description": "Not authorized"
          },
          "403": {
            "description": "Not allowed to access this route"
          }
        }
      }
    }
  }
}